+++ src/battle_anim_effects_1.c
@@ -4921,7 +4921,7 @@ static void UNUSED AnimTask_HideBattlersHealthbox(u8 taskId)
 {
     CMD_ARGS(unk0, unk1);
 
-    u8 i;
+    u32 i;
     for (i = 0; i < gBattlersCount; i++)
     {
         if (cmd->unk0 == TRUE && GetBattlerSide(i) == B_SIDE_PLAYER)
@@ -4936,7 +4936,7 @@ static void UNUSED AnimTask_HideBattlersHealthbox(u8 taskId)
 
 static void UNUSED AnimTask_ShowBattlersHealthbox(u8 taskId)
 {
-    u8 i;
+    u32 i;
     for (i = 0; i < gBattlersCount; i++)
         SetHealthboxSpriteVisible(gHealthboxSpriteIds[i]);
 
@@ -5038,8 +5038,8 @@ static void AnimTask_MoonlightEndFade_Step(u8 taskId)
             u16 color;
             u16 bitmask;
             u16 r3;
-            u16 i;
-            u16 j;
+            u32 i;
+            u32 j;
             task->data[1] = 0;
             if (++task->data[2] <= 15)
             {
@@ -5172,7 +5172,7 @@ static void AnimHornHit_Step(struct Sprite *sprite)
 
 void AnimTask_DoubleTeam(u8 taskId)
 {
-    u16 i;
+    u32 i;
     int obj;
     u16 r3;
     u16 r4;
@@ -5250,8 +5250,8 @@ static void AnimSuperFang(struct Sprite *sprite)
 
 void AnimTask_MusicNotesRainbowBlend(u8 taskId)
 {
-    u16 i;
-    u16 j;
+    u32 i;
+    u32 j;
     u16 index;
 
     index = IndexOfSpritePaletteTag(gParticlesColorBlendTable[0][0]);
@@ -5278,7 +5278,7 @@ void AnimTask_MusicNotesRainbowBlend(u8 taskId)
 // clears the rainbow effect for musical notes.
 void AnimTask_MusicNotesClearRainbowBlend(u8 taskId)
 {
-    u16 i;
+    u32 i;
     for (i = 1; i < ARRAY_COUNT(gParticlesColorBlendTable); i++)
         FreeSpritePaletteByTag(gParticlesColorBlendTable[i][0]);
