+++ src/battle_anim_fire.c
@@ -921,7 +921,7 @@ static void AnimTask_EruptionLaunchRocks_Step(u8 taskId)
 
 static void CreateEruptionLaunchRocks(u8 spriteId, u8 taskId, u8 activeSpritesIdx)
 {
-    u16 i, j;
+    u32 i, j;
     s8 sign;
 
     u16 y = GetEruptionLaunchRockInitialYPos(spriteId);
