+++ src/battle_pyramid_bag.c
@@ -599,7 +599,7 @@ static bool8 LoadPyramidBagGfx(void)
 
 static void SetBagItemsListTemplate(void)
 {
-    u16 i;
+    u32 i;
     u16 *itemIds = gSaveBlock2Ptr->frontier.pyramidBag.itemId[gSaveBlock2Ptr->frontier.lvlMode];
 
     for (i = 0; i < gPyramidBagMenu->listMenuCount - 1; i++)
@@ -767,7 +767,7 @@ static void MovePyramidBagItemSlotInList(u8 from, u8 to)
 
 static void CompactItems(void)
 {
-    u8 i, j;
+    u32 i, j;
     u16 *itemIds = gSaveBlock2Ptr->frontier.pyramidBag.itemId[gSaveBlock2Ptr->frontier.lvlMode];
     u8 *quantities = gSaveBlock2Ptr->frontier.pyramidBag.quantity[gSaveBlock2Ptr->frontier.lvlMode];
 
@@ -791,7 +791,7 @@ static void CompactItems(void)
 
 void UpdatePyramidBagList(void)
 {
-    u16 i;
+    u32 i;
     u16 *itemIds = gSaveBlock2Ptr->frontier.pyramidBag.itemId[gSaveBlock2Ptr->frontier.lvlMode];
 
     CompactItems();
@@ -824,7 +824,7 @@ void UpdatePyramidBagCursorPos(void)
 
 static void InitPyramidBagScroll(void)
 {
-    u8 i;
+    u32 i;
 
     if (gPyramidBagMenuState.cursorPosition > 4)
     {
@@ -1401,7 +1401,7 @@ static void CancelItemSwap(u8 taskId)
 
 void TryStoreHeldItemsInPyramidBag(void)
 {
-    u8 i;
+    u32 i;
     struct Pokemon *party = gPlayerParty;
     u16 *newItems = Alloc(PYRAMID_BAG_ITEMS_COUNT * sizeof(*newItems));
     u8 *newQuantities = Alloc(PYRAMID_BAG_ITEMS_COUNT * sizeof(*newQuantities));
@@ -1434,7 +1434,7 @@ void TryStoreHeldItemsInPyramidBag(void)
 
 static void InitPyramidBagWindows(void)
 {
-    u8 i;
+    u32 i;
 
     InitWindows(sWindowTemplates);
     DeactivateAllTextPrinters();
