+++ src/berry_blender.c
@@ -1544,7 +1544,7 @@ static void SetOpponentsBerryData(u16 playerBerryItemId, u8 playersNum, struct B
     u16 opponentSetId = 0;
     u16 opponentBerryId;
     u16 berryMasterDiff;
-    u16 i;
+    u32 i;
 
     if (playerBerryItemId == ITEM_ENIGMA_BERRY)
     {
@@ -2553,7 +2553,7 @@ static void SendContinuePromptResponse(u16 *cmd)
 
 static void CB2_EndBlenderGame(void)
 {
-    u8 i, j;
+    u32 i, j;
 
     if (sBerryBlender->gameEndState < 3)
         UpdateBlenderCenter();
@@ -3033,7 +3033,7 @@ static void ProcessLinkPlayerCmds(void)
             && sBerryBlender->playerContinueResponses[0] != LINKCMD_BLENDER_STOP
             && sBerryBlender->playerContinueResponses[0] != LINKCMD_SEND_LINK_TYPE)
         {
-            u8 i;
+            u32 i;
 
             // Try to gather responses
             for (i = 0; i < GetLinkPlayerCount(); i++)
@@ -3351,7 +3351,7 @@ static u32 ArrowSpeedToRPM(u16 speed)
 
 static void UpdateRPM(u16 speed)
 {
-    u8 i;
+    u32 i;
     u8 digits[5];
 
     // Check if new max RPM has been reached
@@ -3454,7 +3454,7 @@ static void TryUpdateBerryBlenderRecord(void)
 
 static bool8 PrintBlendingResults(void)
 {
-    u16 i;
+    u32 i;
     s32 xPos, yPos;
 
     struct Pokeblock pokeblock;
@@ -3634,7 +3634,7 @@ static void SortBasedOnPoints(u8 *places, u8 playersNum, u32 *scores)
 static void SortScores(void)
 {
     u8 playerId;
-    u8 i;
+    u32 i;
     u8 places[BLENDER_MAX_PLAYERS];
     u32 points[BLENDER_MAX_PLAYERS];
 
@@ -3664,7 +3664,7 @@ static void SortScores(void)
 
 static bool8 PrintBlendingRanking(void)
 {
-    u16 i;
+    u32 i;
     s32 xPos, yPos;
 
     switch (sBerryBlender->mainState)
